Attempt at eliminating the need for special generated array marshallers.
This is SO CLOSE to working: all we would need is a way of getting the GWT typeid of the array
we want to create. Then we can use JSNI to call Array.initDims() and we're done!
In that case, we would also want to remove all the code and logic responsible for generating
array marshallers. This would shrink both the javascript payload as well as Errai's marshaller
generator complexity.