Switched to recording lifecycle events in a static list. This uncovered a bug in the ID generators, which is now fixed.
The bug was that the when the ID generators probe to see if a given ID value is already in use,
and it is, this causes the entity using that ID to be loaded into the persistence context and
its PostLoad callback to fire. Now the ID generator probes don't cause these side effects.

I also fixed a different bug in the BigInteger ID generator: it was incrementing the ID before returning it,
which could lead to an ID collision if for example it finds ID 123 to be free, but ID 124 is in use.