Added an EventTranslator to refire JPA lifecycle events as CDI events.
This gets rid of the old "yuck" comments where the model knew about the view, and the view
had to look itself up through a static reference. There is still some "yuck" lying around,
but the hope is to modify ErraiEntityManager to fire these events on its own eventually.

Also introduced the new GroceryList type, which will enable removing items from the list.