ERRAI-286 -- Added tests and fixed the feature (which was not fully functional).
The @Service feature implementation is scattered between errai-bus
(ServiceProcessor) and errai-weld-integration (CDIExtensionPoints).
Neither implementation worked fully and because of issues with having
automatic service discovery with errai-cdi on the classpath, it was not
feasible to use both implementations simultaneously.

Consequently, I have re-implemented the missing parts of the feature
(except @Service annotation for fields) within errai-bus.