Fixed bug that caused intermittent failures in EventObserverIntegrationTest
The problem was that we relied on a runtime check (canMarshall) to decided whether or not to advertise a CDI event type to the server. However, this check will always fail for abstract types as canMarshall will return false. So whenever the bus reconnected we failed to advertise observable interfaces and abstract types to the server. Since we already collect all local-only event types on startup this check can now be removed.