Updated JBoss CLI deps to latest (WildFly 8.0.0.Final)
There has been a change in the default management communication protocol: starting
with WildFly 8.0.0.Final, the native management port (9999) is off by default, and
the native management port is deprecated in favour of HTTP management on port
9990.

This update switches to the latest CLI library, but forces the CLI CommandContext
back to native management on port 9999. Max is planning to implement logic that
will try both HTTP on port 9990 and native on port 9999 so we can be compatible
with JBoss AS 7.1.1 through WildFly 8.0.0.Final without any requirement for manual
configuration.