Fixed ERRAI-800: Array marshaller caching causes an error on refresh
The fix is to not cache array marshallers. GWT's rebind cache relies
on the type name provided to the generator which contains the outer
class (the marshaller class that uses the array marshaller). Since
we can't guarantee what this outer class is when lazily generating
array marshallers we can't leverage the built-in rebind cache in
this case.