Changes to conversion behaviour in data-binding (description below).
This introduces two changes to conversion between model and widget types
in Errai Data-Binding:
* Throw an exception if there is no defined conversion between types
(either provided, or a built-in conversion in Convert.to).
* Remove default conversion between java.util.Date and java.lang.String.

As a consequence users must specify a converter when binding a Date
to a String. This is desirable when binding to input elements, since
there input elements of types "date", "datetime", "datetime-local", and "time"
all require different conversion methods.