Only throw exception when unable to convert from String to other type.
A previous change threw an exception in Convert.to(toType, o) if
there was no explicit conversion that could be applied. But this
breaks the scenario where "o" is already assignable to "toType".
Since we can't dynamically check this at runtime, I've added
logging and will only throw exceptions if the object is a String
and there is no built-in conversion to the toType.