Fix issues with DataBinder.setModel (description below).
Fix BindableProxyAgent.syncState to work with input elements and textarea.

Call update widgets in syncState so that if setModel is called with
FROM_UI as the initial state, other widgets bound to model will
be updated.

In BindableProxyAgent.fireChangeEvents, which value is new and which
is old depends on the value of initialState.