First step in data-binding refactoring (description below).
Refactoring data-binding API so that conversion types are locked in
at binding time, and so that converters can be used to enforce
model and widget value types (causing errors if they do not line up).

All conversion between UI and model values is done with converters
and every binding has a non-null converter.