Move CDIAnnotationUtils from errai-ioc (reasoning below).
CDIAnnotationUtils is used by errai-cdi-server in dev mode to
generate the EventQualifierSerializerImpl.class. This means
that SDM can break when errai-ioc is set to provided scoped.

To avoid having to specify a separate profile for SDM, CDIAnnotationUtils
has been moved to errai-codegen. The same problem now exists for errai-codegen
where if it is set to provided scope it may break SDM, however, leaving
errai-codgen as compile scoped and deploying it to the server in production
has less potential for bad side-effects than deploying the errai-ioc jar.