ERRAI-1099: Ability to update the pages state without invoking a full navigation.
This is something that we find very useful when it comes to retaining our pages workflow in the URL. With this we just call updateState with a new set of state data (or a modified version) and we retain the pages initial state.
This introduces the @PageUpdate lifecycle hook that will be invoked when Navigation.updateState is called.