Fix content type issues with the StandardAsyncServlet.
Since the validateContentType method in HttpPollingHandler expects `application/json` we have to ensure our polling is sent as such.
I have a feeling this is actually an issue with the way we are checking inside the HttpPollingHandler. Since I'm not 100% sure how it is meant to work, I will just fix the servlet.